Created validate_file_input for input file paths and sample base names across pipelines.
